Fix auto scrollbar displacing content
After the table of results is populated from first starting to type in
the search field it likely no longer fits vertically on the page. This
caused the browser to provide a scrollbar that would offset the content.

Fixed by using `overflow-y: scroll` to prevent the scrollbar from
appearing and disappearing when the content changes. Instead the
scrollbar is always displayed.